Fixed race condition and state management issues in group video calling (#262)
This commit:

* improves the group call disconnection logic;

* improved the state management of the group call;

* fixes multiple race conditions caused by the asynchronous
  nature of the code.

In particular,  there are some bugs with the implementation of
`RTCPeerConnection` on some browsers. A wrapper class
around this attempts to abtract over these issues so that it can
be used transparently within other code (without having to worry
about the edge cases).

In addition to the above, the wrapper adds some convenience
events, such as `onclose` and `onready`. These merge multiple
internal events into a single one.

Finally, a data channel is added to handle disconnections
gracefully.